requirements.txt
Flask
pymongo
python-dotenv
PyJWT
bcrypt